requirements.txt
anthropic>=0.40.0
icalendar>=5.0.0
flask>=3.0.0
twilio>=9.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
APScheduler>=3.10.0
pytz>=2024.1
